requirements.txt
django
djangorestframework
markdown
django-filter
django_jinja
pyyaml
textile
